requirements.txt
numpy
torch>=1.2
torchvision
matplotlib
tensorboard>=1.15
terminaltables
pillow
tqdm
imgaug>=0.4
wcmatch
omegaconf
loguru
optuna
#wamdb
imgaug
albumentations